+++ sdk/cosmos/azure-cosmos/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 4.2.1 (Unreleased)
-
+**New features**
+- Added language native async i/o client
 
 ## 4.2.0 (2020-10-08)